Get specific about driver support
