+++ examples/arduino/src/main.ino
@@ -1,7 +1,7 @@
 // Define which sensors are attached
 #define SENSOR_ATTACHED_BMP085
 #define SENSOR_ATTACHED_HMC5883L
-#define SENSOR_ATTACHED_MPU6500
+#define SENSOR_ATTACHED_MPU6050
 
 #include <Sensors.h>
 #include <Wire.h>
