+++ examples/mbed/src/main.cpp
@@ -1,7 +1,7 @@
 // Define which sensors are attached
 #define SENSOR_ATTACHED_BMP085
 #define SENSOR_ATTACHED_HMC5883L
-#define SENSOR_ATTACHED_MPU6500
+#define SENSOR_ATTACHED_MPU6050
 
 #include <mbed.h>
 #include <USBSerial.h>
